requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1
Faker==33.1.0
pillow==11.0.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
